Finalizando cadastro cliente

+++ PI/Algoritmos/src/aulas/aula1/cadastroaluno/Aluno.java
@@ -3,21 +3,21 @@ package aulas.aula1.cadastroaluno;
 public class Aluno {
 
     private String nome, ra;
-    private double nota1, nota2, media;
+    private float nota1, nota2, media;
 
-    public double getNota1() {
+    public float getNota1() {
         return nota1;
     }
 
-    public void setNota1(double nota1) {
+    public void setNota1(float nota1) {
         this.nota1 = nota1;
     }
 
-    public double getNota2() {
+    public float getNota2() {
         return nota2;
     }
 
-    public void setNota2(double nota2) {
+    public void setNota2(float nota2) {
         this.nota2 = nota2;
     }
